+++ src/nexgen/beamlines/I19_2_gda_nxs.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 import logging
 from collections import namedtuple
 from pathlib import Path
-from typing import List, Tuple
+from typing import Tuple
 
 import numpy as np
 from numpy.typing import DTypeLike
@@ -26,7 +26,11 @@ from ..nxs_utils.ScanUtils import calculate_scan_points
 from ..nxs_write.NXmxWriter import EventNXmxFileWriter, NXmxFileWriter
 from ..utils import get_iso_timestamp, get_nexus_filename
 from .beamline_parameters import BeamlineAxes
-from .GDAtools.ExtendedRequest import ExtendedRequestIO
+from .GDAtools.ExtendedRequest import (
+    ExtendedRequestIO,
+    read_det_position_from_xml,
+    read_scan_from_xml,
+)
 from .GDAtools.GDAjson2params import JSONParamsIO
 
 # Define a logger object and a formatter
@@ -100,76 +104,6 @@ def log_collection_summary(
 
     logger.info(f"Recorded beam center is: {detector.beam_center}.")
     logger.info(f"Timestamps recorded: {timestamps}")
-
-
-def read_scan_from_xml(ecr: ExtendedRequestIO):
-    """
-    Extract information about the scan contained in the xml file.
-
-    Args:
-        ecr (ExtendedRequestIO): XML tree parser.
-        xmlfile (Path | str): Path to xml file.
-
-    Returns:
-        scan_axis (str): Name of the rotation scan axis
-        pos (Dict): Dictionary containing the (start,end,increment) values for each goniometer axis.
-        num (int): Number of images written.
-    """
-    # Goniometer
-    osc_seq = ecr.getOscillationSequence()
-    # Find scan range
-    if osc_seq["range"] == 0.0:
-        scan_range = (osc_seq["start"], osc_seq["start"])
-        num = osc_seq["number_of_images"]
-    else:
-        start = osc_seq["start"]
-        num = osc_seq["number_of_images"]
-        stop = start + num * osc_seq["range"]
-        scan_range = (start, stop)
-    # Determine scan axis
-    if ecr.getAxisChoice() == "omega":
-        scan_axis = "omega"
-        omega_pos = (*scan_range, osc_seq["range"])  # 0.0)
-        phi_pos = (*2 * (ecr.getOtherAxis(),), 0.0)
-    else:
-        scan_axis = "phi"
-        phi_pos = (*scan_range, osc_seq["range"])  # 0.0)
-        omega_pos = (*2 * (ecr.getOtherAxis(),), 0.0)
-    kappa_pos = (*2 * (ecr.getKappa(),), 0.0)
-
-    pos = {
-        "omega": omega_pos,
-        "phi": phi_pos,
-        "kappa": kappa_pos,
-        "sam_x": (0.0, 0.0, 0.0),
-        "sam_y": (0.0, 0.0, 0.0),
-        "sam_z": (0.0, 0.0, 0.0),
-    }
-
-    return scan_axis, pos, int(num)
-
-
-def read_det_position_from_xml(
-    ecr: ExtendedRequestIO,
-    det_description: str,
-) -> List[float]:
-    """Extract the detector position contained in the xml file.
-
-    Args:
-        ecr (ExtendedRequestIO): XML tree parser.
-        det_description (str): Detector description
-
-    Returns:
-        List[float]: Detector axes positions in the order [2theta, det_z]
-    """
-    if "tristan" in det_description.lower():
-        return [0.0, ecr.getSampleDetectorDistance()]
-    elif "eiger" in det_description.lower():
-        return [ecr.getTwoTheta(), ecr.getSampleDetectorDistance()]
-    logger.warning(
-        "Unrecognized detector description. Setting two_theta to 0.0 to be safe."
-    )
-    return [0.0, ecr.getSampleDetectorDistance()]
 
 
 def tristan_writer(
